server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Amal <amal.mukhtar@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
pyjwt = "^2.6.0"
django-cors-headers = "^3.13.0"
python-dateutil = "^2.8.2"


[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
